requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
aniso8601==9.0.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
bidict==0.22.1
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
coloredlogs==15.0.1
dnspython==2.4.2
email-validator==1.1.3
eventlet==0.33.3
fastapi==0.99.1
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.0
gprofiler-official==1.0.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
htmlmin==0.1.12
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsmin==3.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.3
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mygene==3.2.2
networkx==3.1
nltk==3.8.1
numpy==1.25.2
onnxruntime==1.15.1
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
platformdirs==3.10.0
ply==3.11
pooch==1.7.0
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pydantic==1.10.12
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.3
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
simple-websocket==0.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.29
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.0
xxhash==3.3.0
yarl==1.9.2
